Starting to refactor framework for 'persisting' Knowledgegraph state across pages
